fix: code-refactor

+++ tests/v1/blog/delete_blog_test.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 from uuid_extensions import uuid7
 
 from api.db.database import get_db
-from api.utils.dependencies import get_super_admin
+from api.v1.services.user import user_service
 from api.v1.models import User
 from api.v1.models.blog import Blog
 from main import app
@@ -17,7 +17,7 @@ def mock_get_db():
     yield db_session
 
 
-def mock_get_super_admin():
+def mock_get_current_super_admin():
     return User(id="1", is_super_admin=True)
 
 @pytest.fixture
@@ -32,52 +32,53 @@ def client(db_session_mock):
     yield client
     
 
-
-
-
-
 def test_delete_blog_success(client, db_session_mock):
+    '''Test for success in blog deletion'''
+
     app.dependency_overrides[get_db] = lambda: db_session_mock
-    app.dependency_overrides[get_super_admin] = lambda: mock_get_super_admin
+    app.dependency_overrides[user_service.get_current_super_admin] = lambda: mock_get_current_super_admin
     blog_id = uuid7()
     mock_blog = Blog(id=blog_id, title="Test Blog",
                      content="Test Content", is_deleted=False)
 
     db_session_mock.query(Blog).filter(id==blog_id).first.return_value.id = [mock_blog]
 
-    response = client.delete(f"/api/v1/blogs/{mock_blog.id}")
+    response = client.delete(f"/api/v1/blogs/{mock_blog.id}", headers={'Authorization': 'Bearer token'})
 
-
-    assert response.status_code == 200
-    assert response.json() == {
-        "message": "Blog post deleted successfully", "status_code": 200}
+    assert response.status_code == 204
 
 
 def test_delete_blog_unauthorized(client, db_session_mock):
-    blog_id = uuid7()
-    mock_blog = Blog(id=blog_id, title="Test Blog",
-                     content="Test Content", is_deleted=False)
-    app.dependency_overrides[get_super_admin] = lambda: None
+    '''Test for unauthorized user'''
 
-    response = client.delete(f"/api/v1/blogs/{mock_blog.id}")
+    blog_id = str(uuid7())
+    mock_blog = Blog(
+        id=blog_id, 
+        title="Test Blog",
+        content="Test Content", 
+        is_deleted=False
+    )
+    
+    app.dependency_overrides[user_service.get_current_super_admin] = lambda: None
 
+    response = client.delete(f"/api/v1/blogs/{mock_blog.id}", headers={'Authorization': 'Bearer token'})
 
-    assert response.json()["status_code"] == 403
-    assert response.json()["message"] == "Unauthorized User"
+    assert response.status_code == 401
 
     
 def test_delete_blog_not_found(client, db_session_mock):
+    '''test for blog not found'''
+
     db_session_mock.query(Blog).filter(Blog.id == f'{uuid7()}').first.return_value = None
 
     app.dependency_overrides[get_db] = lambda: db_session_mock
-    app.dependency_overrides[get_super_admin] = lambda: mock_get_super_admin
+    app.dependency_overrides[user_service.get_current_super_admin] = lambda: mock_get_current_super_admin
 
-    response = client.delete(f"api/v1/blogs/{uuid7()}")
+    response = client.delete(f"api/v1/blogs/{uuid7()}", headers={'Authorization': 'Bearer token'})
 
     assert response.json()["status_code"] == 404
-    assert response.json()["message"] == "Blog with the given ID does not exist"
+    assert response.json()["message"] == "Post not found"
     
-
 
 if __name__ == "__main__":
     pytest.main()
